+++ CHANGELOG.md
@@ -3,12 +3,14 @@
 ## Update
 
 - [PoolDetail] Improve loading behavior [#2234](https://github.com/thorchain/asgardex-electron/issues/2234)
+- [Swap] Change default slippage tolerance to 3% [#2237](https://github.com/thorchain/asgardex-electron/issues/2237)
 
 ## Fix
 
 - White screen with Ledger Terra wallet connected (#2227)[https://github.com/thorchain/asgardex-electron/issues/2227]
 - [Wallet] Locking wallet at Deposit or Withdraw failed (#2233)[https://github.com/thorchain/asgardex-electron/issues/2233]
 - [e2e] Testcafe is failing [#750](https://github.com/thorchain/asgardex-electron/issues/750)
+- [Swap] Consider fees in swap limit [#2243](https://github.com/thorchain/asgardex-electron/pull/2242)
 
 ## Internal